requirements.txt
# JBridge Invoice Processor Dependencies
google-generativeai>=0.8.3
PyPDF2>=3.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0

# Original DocGenius dependencies (optional, for PDFChat)
tiktoken
faiss-cpu
langchain
langchain-community
langchain-openai
